+++ CHANGELOG.rst
@@ -4,7 +4,8 @@ CHANGELOG
 6.3.11 (unreleased)
 -------------------
 
-- Nothing changed yet.
+- Fix order by int and float Search pgcatalog
+  [rboixaderg]
 
 
 6.3.10 (2021-06-07)